[feature] Search: implement API endpoint and Dto
