+++ app/src/main/java/com/rwmobi/giphytrending/data/source/network/GiphyApi.kt
@@ -5,9 +5,12 @@
 
 package com.rwmobi.giphytrending.data.source.network
 
+import com.rwmobi.giphytrending.data.source.network.dto.SearchNetworkResponseDto
 import com.rwmobi.giphytrending.data.source.network.dto.TrendingNetworkResponseDto
 import retrofit2.http.GET
 import retrofit2.http.Query
+
+// https://developers.giphy.com/docs/api/endpoint#trending
 
 interface GiphyApi {
     @GET("v1/gifs/trending")
@@ -17,4 +20,13 @@ interface GiphyApi {
         @Query("offset") offset: Int,
         @Query("rating") rating: String,
     ): TrendingNetworkResponseDto
+
+    @GET("v1/gifs/search")
+    suspend fun getSearch(
+        @Query("api_key") apiKey: String,
+        @Query("q") keyword: String,
+        @Query("limit") limit: Int, // ⚠️ observe API limit
+        @Query("offset") offset: Int,
+        @Query("rating") rating: String,
+    ): SearchNetworkResponseDto
 }
