+++ app/src/main/java/com/rwmobi/giphytrending/data/source/network/dto/SearchNetworkResponseDto.kt
@@ -1,0 +1,19 @@
+/*
+ * Copyright (c) 2024. Ryan Wong
+ * https://github.com/ryanw-mobile
+ */
+
+package com.rwmobi.giphytrending.data.source.network.dto
+
+import androidx.annotation.Keep
+import com.squareup.moshi.Json
+
+@Keep
+data class SearchNetworkResponseDto(
+    @Json(name = "data")
+    val trendingData: List<TrendingDataDto>,
+    @Json(name = "meta")
+    val metaDto: MetaDto,
+    @Json(name = "pagination")
+    val pagination: PaginationDto,
+)
